# core-lending-regulatory-compliance-interfaces/src/main/java/com/firefly/core/lending/compliance/interfaces/dtos/reporting/v1/RegulatorySubmissionDTO.java
/*
 * Copyright 2025 Firefly Software Foundation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
 * You may obtain a copy of the License at
 *
 *     http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing, software
 * distributed under the License is distributed on an "AS IS" BASIS,
 * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 * See the License for the specific language governing permissions and
 * limitations under the License.
 */


package com.firefly.core.lending.compliance.interfaces.dtos.reporting.v1;

import com.firefly.core.lending.compliance.interfaces.enums.reporting.v1.SubmissionStatusEnum;
import org.fireflyframework.utils.annotations.FilterableId;
import com.fasterxml.jackson.annotation.JsonProperty;
import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.NotBlank;
import jakarta.validation.constraints.Size;
import jakarta.validation.constraints.PastOrPresent;
import java.time.LocalDate;
import java.time.LocalDateTime;
import java.util.UUID;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class RegulatorySubmissionDTO {
    @JsonProperty(access = JsonProperty.Access.READ_ONLY)
    private UUID regulatorySubmissionId;

    @FilterableId
    @NotNull(message = "Reporting run ID is required")
    private UUID reportingRunId;

    @NotBlank(message = "Submission reference is required")
    @Size(max = 100, message = "Submission reference cannot exceed 100 characters")
    private String submissionRef;

    @NotNull(message = "Submission date is required")
    @PastOrPresent(message = "Submission date cannot be in the future")
    private LocalDate submissionDate;

    @NotNull(message = "Submission status is required")
    private SubmissionStatusEnum submissionStatus;

    @Size(max = 2000, message = "Response details cannot exceed 2000 characters")
    private String responseDetails;

    private LocalDateTime createdAt;
    private LocalDateTime updatedAt;
}