# core-lending-regulatory-compliance-sdk/target/classes/api-spec/openapi.yml
openapi: 3.0.1
info:
  title: core-lending-regulatory-compliance
  description: Lending Regulatory & Compliance Core Application
  contact:
    name: Firefly Software Foundation
    email: dev@getfirefly.io
  version: 1.0.0
servers:
  - url: http://core.getfirefly.io/loan-compliance
    description: Development Environment
  - url: /
    description: Local Development Environment
...
paths:
...
  /api/v1/reporting-runs/{reportingRunId}/submissions/{regulatorySubmissionId}:
    get:
      tags:
        - RegulatorySubmission
      summary: Get a regulatory submission by ID
      operationId: getById_1
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: regulatorySubmissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RegulatorySubmissionDTO'
    put:
      tags:
        - RegulatorySubmission
      summary: Update a regulatory submission
      operationId: update_1
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: regulatorySubmissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegulatorySubmissionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RegulatorySubmissionDTO'
    delete:
      tags:
        - RegulatorySubmission
      summary: Delete a regulatory submission record
      operationId: delete_1
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: regulatorySubmissionId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/aml-cases/{amlCaseId}/sars/{amlSarId}:
    get:
      tags:
        - AmlSar
      summary: Get a SAR by ID
      operationId: getById_4
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amlSarId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlSarDTO'
    put:
      tags:
        - AmlSar
      summary: Update a SAR record
      operationId: update_4
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amlSarId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlSarDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlSarDTO'
    delete:
      tags:
        - AmlSar
      summary: Delete a SAR record
      operationId: delete_4
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: amlSarId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: OK
...
  /api/v1/reporting-runs/{reportingRunId}/submissions:
    get:
      tags:
        - RegulatorySubmission
      summary: List or search regulatory submissions for a specific run
      operationId: findAll_1
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.reportingRunId
          in: query
          description: Exact filter for reporting run id
          required: false
          schema:
            type: string
        - name: filters.submissionRef
          in: query
          description: Exact filter for submission ref
          required: false
          schema:
            type: string
        - name: filters.submissionDate
          in: query
          description: Exact filter for submission date
          required: false
          schema:
            type: string
        - name: filters.submissionStatus
          in: query
          description: Exact filter for submission status
          required: false
          schema:
            type: string
            enum:
              - SUBMITTED
              - ACCEPTED
              - REJECTED
              - REVISED
              - SUPERSEDED
        - name: filters.responseDetails
          in: query
          description: Exact filter for response details
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseRegulatorySubmissionDTO'
    post:
      tags:
        - RegulatorySubmission
      summary: Create a new regulatory submission
      operationId: create_1
      parameters:
        - name: reportingRunId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegulatorySubmissionDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/RegulatorySubmissionDTO'
...
  /api/v1/aml-cases/{amlCaseId}/sars:
    get:
      tags:
        - AmlSar
      summary: List or search SARs for a given AML case
      operationId: findAll_4
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: pagination.pageNumber
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            format: int32
            default: 0
        - name: pagination.pageSize
          in: query
          description: Number of items per page
          required: false
          schema:
            type: integer
            format: int32
            default: 10
        - name: pagination.sortBy
          in: query
          description: Field to sort by
          required: false
          schema:
            type: string
        - name: pagination.sortDirection
          in: query
          description: Sort direction (ASC or DESC)
          required: false
          schema:
            type: string
            default: DESC
        - name: options.caseInsensitiveStrings
          in: query
          description: Enable case-insensitive string filtering
          required: false
          schema:
            type: boolean
            default: false
        - name: options.includeInheritedFields
          in: query
          description: Include fields from parent classes
          required: false
          schema:
            type: boolean
            default: false
        - name: filters.amlCaseId
          in: query
          description: Exact filter for aml case id
          required: false
          schema:
            type: string
        - name: filters.sarReference
          in: query
          description: Exact filter for sar reference
          required: false
          schema:
            type: string
        - name: filters.sarDate
          in: query
          description: Exact filter for sar date
          required: false
          schema:
            type: string
        - name: filters.sarStatus
          in: query
          description: Exact filter for sar status
          required: false
          schema:
            type: string
            enum:
              - DRAFT
              - SUBMITTED
              - FEEDBACK_RECEIVED
              - CLOSED
              - AMENDED
        - name: filters.authorityFeedback
          in: query
          description: Exact filter for authority feedback
          required: false
          schema:
            type: string
        - name: filters.createdAt
          in: query
          description: Exact filter for created at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].from
          in: query
          description: Filter created at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[createdAt].to
          in: query
          description: Filter created at to value
          required: false
          schema:
            type: string
            format: date-time
        - name: filters.updatedAt
          in: query
          description: Exact filter for updated at
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].from
          in: query
          description: Filter updated at from value
          required: false
          schema:
            type: string
            format: date-time
        - name: rangeFilters.ranges[updatedAt].to
          in: query
          description: Filter updated at to value
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/PaginationResponseAmlSarDTO'
    post:
      tags:
        - AmlSar
      summary: Create a new SAR (Suspicious Activity Report)
      operationId: create_4
      parameters:
        - name: amlCaseId
          in: path
          required: true
          schema:
            type: string
            format: uuid
        - name: X-Idempotency-Key
          in: header
          description: Unique key for idempotent requests. If provided, ensures that identical requests with the same key will only be processed once.
          required: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AmlSarDTO'
        required: true
      responses:
        '200':
          description: OK
          content:
            '*/*':
              schema:
                $ref: '#/components/schemas/AmlSarDTO'
...
components:
  schemas:
...
    RegulatorySubmissionDTO:
      required:
        - reportingRunId
        - submissionDate
        - submissionRef
        - submissionStatus
      type: object
      properties:
        regulatorySubmissionId:
          type: string
          format: uuid
          readOnly: true
        reportingRunId:
          type: string
          format: uuid
        submissionRef:
          maxLength: 100
          minLength: 0
          type: string
        submissionDate:
          type: string
          format: date
        submissionStatus:
          type: string
          enum:
            - SUBMITTED
            - ACCEPTED
            - REJECTED
            - REVISED
            - SUPERSEDED
        responseDetails:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    AmlSarDTO:
      required:
        - amlCaseId
        - sarDate
        - sarReference
        - sarStatus
      type: object
      properties:
        amlSarId:
          type: string
          format: uuid
          readOnly: true
        amlCaseId:
          type: string
          format: uuid
        sarReference:
          maxLength: 100
          minLength: 0
          type: string
        sarDate:
          type: string
          format: date
        sarStatus:
          type: string
          enum:
            - DRAFT
            - SUBMITTED
            - FEEDBACK_RECEIVED
            - CLOSED
            - AMENDED
        authorityFeedback:
          maxLength: 2000
          minLength: 0
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
...
    PaginationResponseRegulatorySubmissionDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/RegulatorySubmissionDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...
    PaginationResponseAmlSarDTO:
      type: object
      properties:
        content:
          type: array
          description: The list of items for the current page.
          items:
            $ref: '#/components/schemas/AmlSarDTO'
        totalElements:
          type: integer
          description: The total number of elements across all pages.
          format: int64
        totalPages:
          type: integer
          description: The total number of pages based on the data size and page size.
          format: int32
        currentPage:
          type: integer
          description: The current page number, typically zero-based.
          format: int32
      description: Represents a paginated response containing a list of items and pagination metadata.
...